Remove Prometheum Attribute Event, and simplify Prometheum handling
This event did not actually function.
Tags are not populated until the world loads,
which is way too late for this event.

The tag checks for the Prometheum tick are
also not strictly needed to handle everything correctly.
The handling already requires armor to be an ArmorItem,
and tools to hold attributes.