Add system for handling custom attributes on vanilla tool types
This is not a great system, and needs to be refactored at some point.
I believe parts of this is solved in the latest versions, although in this version the attribute handling for custom tools is a major pain point.
You either reimplement the tool logic from scratch, commit gross hacks into the ToolMaterial **record**, or you disgustingly abuse the postProcessComponent method which I've done.
I don't think it should have a noticeable performance impact due to the JVM optimizing this, but this method *will* become a hotspot checking for empty lists.
Hopefully this vanishes once I port to latest...