Fix #307
Oversight with the BlockSet refactor for 1.21.4.
Specifically block sets were all using the variant name
for registering their blocks.
This is what the game uses to determine which loot table
to process.
Therefore, a "deepslate" variant would look for the "deepslate_ore" loot table.
This has been fixed by explicitly naming the parameter to `variantName`,
which should prevent the confusion.